requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
matplotlib>=3.5.1
pillow>=9.0.1
opencv-python>=4.5.5
scikit-learn>=1.0.2
tk>=0.1.0